修复bug，更新数据库备份

+++ dgutshop-core/src/main/java/org/dgutstu/dgutshop/core/security/JwtAuthenticationTokenFilter.java
@@ -80,7 +80,7 @@ public class JwtAuthenticationTokenFilter extends BasicAuthenticationFilter {
                         //组装参数
                         LoginUser loginUser = new LoginUser();
                         loginUser.setUsername(claims.getSubject());
-                        loginUser.setUserId(Long.parseLong(claims.getId()));
+                        loginUser.setUserId(Integer.parseInt(claims.getId()));
                         loginUser.setAuthorities(authorities);
 //                    System.out.println(loginUser);
                         UsernamePasswordAuthenticationToken authentication = new UsernamePasswordAuthenticationToken(loginUser, userId, authorities);
